实现Jwt+Redis双验证顶号踢人，管理禁用踹人下线

+++ laolao-pc-builds/src/main/java/com/laolao/mapper/common/UserCommonMapper.java
@@ -16,7 +16,7 @@ public interface UserCommonMapper {
 
 
     @Select("select * from user where id = #{userId}")
-    User getUser(long userId);
+    User getUser(int userId);
 
     List<User> selectUser(List<Integer> userIdList1);
